+++ src/types.ts
@@ -390,3 +390,4 @@ declare const process: {
 }
 
 export { Logger }
+
