+++ src/waka-realtime/BaseRealtime.ts
@@ -1,15 +1,10 @@
-import path from 'path'
 import { AxiosInstance } from 'axios'
+import path from 'path'
 import { oc } from 'ts-optchain'
-import WakaRedis from './Redis'
-import { Logger } from '../typings'
-import {
-  VehiclePosition,
-  AlertFeedEntity,
-  PositionFeedEntity,
-  UpdateFeedEntity,
-} from '../gtfs'
+import { AlertFeedEntity, PositionFeedEntity, UpdateFeedEntity, VehiclePosition } from '../gtfs'
+import { Logger } from '../types'
 import { check } from '../utils'
+import WakaRedis from './Redis'
 
 export const PROTOBUF_PATH = [
   path.join(__dirname, 'protos', 'gtfs-realtime.proto'),
