+++ src/waka-worker/index.ts
@@ -1,17 +1,16 @@
 import { Router } from 'express'
-import * as morgan from 'morgan'
-import createLogger from './logger'
 import cityMetadataJSON from '../cityMetadata.json'
-import Connection from './db/connection'
-import Lines from './lines'
-import Search from './stops/search'
-import Station from './stops/station'
-import StopsNZAKL from './stops/regions/nz-akl'
-import StopsNZWLG from './stops/regions/nz-wlg'
-import Realtime from './realtime'
-import { WorkerConfig, Logger } from '../typings'
+import { Logger, WorkerConfig } from '../types'
 import BaseStops from '../types/BaseStops'
 import WakaRedis from '../waka-realtime/Redis'
+import Connection from './db/connection'
+import Lines from './lines'
+import createLogger from './logger'
+import Realtime from './realtime'
+import StopsNZAKL from './stops/regions/nz-akl'
+import StopsNZWLG from './stops/regions/nz-wlg'
+import Search from './stops/search'
+import Station from './stops/station'
 
 class WakaWorker {
   config: WorkerConfig
