+++ src/waka-worker/stops/search.ts
@@ -1,10 +1,10 @@
-import * as sql from 'mssql'
 import * as Logger from 'bunyan'
 import { Response } from 'express'
+import * as sql from 'mssql'
 import { oc } from 'ts-optchain'
-import Connection from '../db/connection'
-import { WakaRequest, StopRouteType } from '../../typings'
+import { StopRouteType, WakaRequest } from '../../types'
 import BaseStops from '../../types/BaseStops'
+import Connection from '../db/connection'
 
 interface SearchProps {
   logger: Logger
